pyproject.toml
[tool.poetry]
name = "qontract-development-cli"
version = "0.5.0"
description = "Helper tool for qontract-reconcile development"
authors = ["Christian Assing <cassing@redhat.com>"]
license = "MIT"
readme = "README.md"
repository = "http://github.com/chassing/qontract-development-cli"
homepage = "http://github.com/chassing/qontract-development-cli"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
Jinja2 = "^3.1.2"
typer = "^0.6.1"
rich = "^12.5.1"
pydantic = "^1.10.2"
appdirs = "^1.4.4"
PyYAML = "^6.0"
getkey = "^0.6.5"
watchfiles = "^0.18.0"
textual = "^0.2.1"

[tool.poetry.dev-dependencies]
types-PyYAML = "^6.0.12"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
mypy = "^0.982"
pytest = "^7.2.0"
commitizen = "^2.37.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
qd = "qontract_development_cli.__main__:app"

[tool.mypy]
plugins = "pydantic.mypy"
enable_error_code = "truthy-bool, redundant-expr"
no_implicit_optional = true
check_untyped_defs = true
warn_unused_ignores = true
show_error_codes = true


[[tool.mypy.overrides]]
module = ["getkey.*"]
ignore_missing_imports = true

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.5.0"
version_files = ["pyproject.toml:version"]
update_changelog_on_bump = true
major_version_zero = true
change_type_map = { "feat" = "Features", "fix" = "Bug Fixes" }
tag_format = "v$version"

[tool.isort]
profile = "black"
multi_line_output = 3
force_grid_wrap = 2
float_to_top = true
